Convert sliders/dials to front/backend
